fix paginated data load
